+++ packages/credential-sdk/src/types/public-keys/public-key-sr25519-value.js
@@ -1,11 +1,1 @@
-import { sized, TypedBytes } from '../generic';
-import { Sr25519VerKeyName } from '../../vc/crypto/constants';
-
-/** Class representing value of a Sr25519 PublicKey */
-export default class PublicKeySr25519Value extends sized(TypedBytes) {
-  static Type = 'sr25519';
-
-  static Size = 32;
-
-  static VerKeyType = Sr25519VerKeyName;
-}
+export { PublicKeySr25519Value as default } from '@docknetwork/crypto-utils/types';