+++ packages/credential-sdk/src/types/public-keys/public-key-x25519-value.js
@@ -1,10 +1,1 @@
-import { sized, TypedBytes } from '../generic';
-
-/** Class representing value of a X25519 PublicKey */
-export default class PublicKeyX25519Value extends sized(TypedBytes) {
-  static Type = 'x25519';
-
-  static Size = 32;
-
-  static VerKeyType = 'X25519KeyAgreementKey2019';
-}
+export { PublicKeyX25519Value as default } from '@docknetwork/crypto-utils/types';