+++ packages/credential-sdk/src/types/signatures/signature-secp256k1-value.js
@@ -1,9 +1,1 @@
-import SignatureValue from './signature-value';
-import { PublicKeySecp256k1 } from '../public-keys';
-
-/** Class representing a Secp256k1 Signature */
-export default class SignatureSecp256k1Value extends SignatureValue {
-  static PublicKey = PublicKeySecp256k1;
-
-  static Size = 65;
-}
+export { SignatureSecp256k1Value as default } from '@docknetwork/crypto-utils/types';