+++ packages/credential-sdk/src/types/signatures/signature-sr25519-value.js
@@ -1,9 +1,1 @@
-import { PublicKeySr25519 } from '../public-keys';
-import SignatureValue from './signature-value';
-
-/** Class representing a Sr25519 Signature */
-export default class SignatureSr25519Value extends SignatureValue {
-  static PublicKey = PublicKeySr25519;
-
-  static Size = 64;
-}
+export { SignatureSr25519Value as default } from '@docknetwork/crypto-utils/types';