+++ packages/crypto-utils/src/index.js
@@ -1,0 +1,4 @@
+export * from './keypairs';
+export * from './types';
+export * from './utils';
+export * from './vc';