+++ packages/crypto-utils/src/types/public-keys/public-key-x25519-value.js
@@ -1,0 +1,10 @@
+import { sized, TypedBytes } from '../generic';
+
+/** Class representing value of a X25519 PublicKey */
+export default class PublicKeyX25519Value extends sized(TypedBytes) {
+  static Type = 'x25519';
+
+  static Size = 32;
+
+  static VerKeyType = 'X25519KeyAgreementKey2019';
+}